extract resource collection
