+++ src/Illuminate/Http/Resource.php
@@ -7,7 +7,6 @@ use ArrayAccess;
 use ArrayIterator;
 use JsonSerializable;
 use IteratorAggregate;
-use Illuminate\Support\Str;
 use Illuminate\Support\Collection;
 use Illuminate\Container\Container;
 use Illuminate\Pagination\AbstractPaginator;
@@ -74,29 +73,6 @@ class Resource implements ArrayAccess, IteratorAggregate, JsonSerializable, Resp
     public function __construct($resource)
     {
         $this->resource = $resource;
-
-        if ($this->isCollectionResource()) {
-            $this->resource = $this->collectResource($resource);
-        }
-    }
-
-    /**
-     * Map the given collection resource into its individual resources.
-     *
-     * @param  mixed  $resource
-     * @return mixed
-     */
-    protected function collectResource($resource)
-    {
-        if (! $this->collects) {
-            throw new Exception('The ['.get_class($this).'] resource must specify the models it collects.');
-        }
-
-        $this->collection = $resource->mapInto($this->collects);
-
-        return $resource instanceof Collection
-                    ? $this->collection
-                    : $resource->setCollection($this->collection);
     }
 
     /**
@@ -154,9 +130,7 @@ class Resource implements ArrayAccess, IteratorAggregate, JsonSerializable, Resp
      */
     public function json()
     {
-        return $this->resource instanceof AbstractPaginator
-                    ? new Resources\PaginatedJsonResourceResponse($this)
-                    : new Resources\JsonResourceResponse($this);
+        return new Resources\JsonResourceResponse($this);
     }
 
     /**
@@ -167,35 +141,7 @@ class Resource implements ArrayAccess, IteratorAggregate, JsonSerializable, Resp
      */
     public function toJson($request)
     {
-        return $this->isCollectionResource()
-                    ? $this->collectionToJson($request)
-                    : $this->resourceToJson($request);
-    }
-
-    /**
-     * Determine if this resource is a collection resource.
-     *
-     * @return bool
-     */
-    protected function isCollectionResource()
-    {
-        return $this->resource instanceof Collection ||
-               $this->resource instanceof AbstractPaginator;
-    }
-
-    /**
-     * Convert the collection into a JSON array.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array
-     */
-    public function collectionToJson($request)
-    {
-        $data = $this->resource->map(function ($item) use ($request) {
-            return $item->toJson($request);
-        })->all();
-
-        return static::$wrap ? [static::$wrap => $data] : $data;
+        return $this->resourceToJson($request);
     }
 
     /**
